Add files via upload
These files are a starting point for use and development of a molecular barcoding lab activity. Custom tailoring to your class will be required. The molecular barcoding supplies plant fungal concentration.xlsx was created to equip several classrooms simultaneously carrying out molecular barcoding. Pricing for individual supplies will vary but this is a fairly complete list of equipment needed (we think). Special thanks to MES and BP for their significant contributions and sharing of their time and efforts.